+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changes to be released in next version
  * 
 
 🙌 Improvements
- * 
+ * MXCrypto: Introduce MXCryptoVersion to track changes between MXCrypto module updates.
 
 🐛 Bugfix
  * MXOlmDevice: Make usage of libolm data process-safe (vector-im/element-ios/3817).